Remove python3 calls or convert to generic python calls.
Remove all of the package "/usr/bin/env python3" lines at
the tops of the files, since these are never executed
directly anyway. Also, convert the "python3" calls
for the top-level "ur" script and for the setup.py
script, so that they can work with either flavor of
python, as the system see fit.